Restructure python package
